Padronizar nome das classes
