+++ jbugbrasil-telegram-bot/plugins/karma-plugin/src/main/resources/META-INF/services/br.com.jbugbrasil.bot.api.spi.CommandProvider
@@ -1,1 +1,1 @@
-br.com.jbugbrasil.bot.service.karma.KarmaCommandProvider
+br.com.jbugbrasil.bot.service.karma.KarmaCommand
