+++ jbugbrasil-telegram-bot/plugins/karma-plugin/src/main/resources/META-INF/services/br.com.jbugbrasil.bot.api.spi.PluginProvider
@@ -1,1 +1,1 @@
-br.com.jbugbrasil.bot.plugin.karma.KarmaPluginProvider
+br.com.jbugbrasil.bot.plugin.karma.KarmaPlugin
