+++ jbugbrasil-telegram-bot/services/faq-service/src/main/resources/META-INF/services/br.com.jbugbrasil.bot.api.spi.CommandProvider
@@ -1,1 +1,1 @@
-br.com.jbugbrasil.bot.service.faq.FaqCommandProvider
+br.com.jbugbrasil.bot.service.faq.FaqCommand
