+++ jbugbrasil-telegram-bot/services/jboss-books-service/src/main/resources/META-INF/services/br.com.jbugbrasil.bot.api.spi.CommandProvider
@@ -1,1 +1,1 @@
-br.com.jbugbrasil.bot.service.jbossbooks.command.JBossBooksCommandProvider
+br.com.jbugbrasil.bot.service.jbossbooks.command.JBossBooks
