+++ jbugbrasil-telegram-bot/services/ping-service/src/main/resources/META-INF/services/br.com.jbugbrasil.bot.api.spi.CommandProvider
@@ -1,1 +1,1 @@
-br.com.jbugbrasil.bot.service.ping.PingCommandProvider
+br.com.jbugbrasil.bot.service.ping.Ping
