+++ jbugbrasil-telegram-bot/services/uptime-service/src/main/resources/META-INF/services/br.com.jbugbrasil.bot.api.spi.CommandProvider
@@ -1,1 +1,1 @@
-br.com.jbugbrasil.bot.service.uptime.UptimeCommandProvider
+br.com.jbugbrasil.bot.service.uptime.Uptime
