+++ jbugbrasil-telegram-bot/services/urban-dictionary-service/src/main/resources/META-INF/services/br.com.jbugbrasil.bot.api.spi.CommandProvider
@@ -1,1 +1,1 @@
-br.com.jbugbrasil.bot.service.urbandictionary.UrbanDictionaryCommandProvider
+br.com.jbugbrasil.bot.service.urbandictionary.UrbanDictionary
