+++ jbugbrasil-telegram-bot/services/weather-service/src/main/resources/META-INF/services/br.com.jbugbrasil.bot.api.spi.CommandProvider
@@ -1,1 +1,1 @@
-br.com.jbugbrasil.bot.service.weather.WeatherService
+br.com.jbugbrasil.bot.service.weather.Weather
